some cleanup, optimized timings
Cleaned up the debug messages, had some spelling errors and unit errors in them.
Added a delay(100) after the ESP.deepSleep instructions to ensure they definitely get executed.
Changed the wifi connect check behavior so he now checks every 0.5 s if he is connected and kills the process after 5 s. This is to speed up the process, if the timing is troublesome, just change the attempts to double the amount and give him 10s.