replace files with links, made serial debug optional, added send_webhook() error handling
Replaced the libary files with links to our submodules to keep things in sync.
Added preprocessor define SERIAL_DEBUG to be able to enable/disable the serial debug messages.
Added the code to catch the "false" return of send_webhook() I patched into it, and process it so we may try again in case of error.
Corrected some comments on the retry counter of wifi connecting routine.